Move font-size-related functions out of Termonad.App
